\docType{package}
\name{polywog-package}
\alias{polywog-package}
\title{Bootstrapped basis regression with oracle model selection}
\description{
  A package for flexible functional form estimation via
  bootstrapped basis regression with oracle model
  selection.  This version of the software should be
  considered \strong{in beta}.  Please email the
  maintainer, Brenton Kenkel
  (\email{brenton.kenkel@gmail.com}), with bug reports or
  feature requests.
}
\section{Acknowledgements}{
  We thank the Wallis Institute of Political Economy for
  financial support.
}
\references{
  Brenton Kenkel and Curtis S. Signorino.  2012.  "A Method
  for Flexible Functional Form Estimation: Bootstrapped
  Basis Regression with Variable Selection."  Typescript,
  University of Rochester.
}

