% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polywog.r
\docType{package}
\name{polywog-package}
\alias{polywog-package}
\title{Bootstrapped basis regression with oracle model selection}
\description{
A package for flexible functional form estimation via bootstrapped basis
regression with oracle model selection.  This version of the software should
be considered \strong{in beta}.  For bug reports and feature requests,
please email Brenton Kenkel (\email{brenton.kenkel@gmail.com}) or file an
issue at \url{https://github.com/brentonk/polywog-package/issues}.
}
\section{Acknowledgements}{
 We are grateful to Tyson Chatangier for many
helpful suggestions about an earlier version of the package.  We also thank
the Wallis Institute of Political Economy and the Theory and Statistics
Research Lab at the University of Rochester for financial support during the
writing of the package.
}

\references{
Brenton Kenkel and Curtis S. Signorino.  2012.  "A Method for Flexible
Functional Form Estimation: Bootstrapped Basis Regression with Variable
Selection."  Typescript, University of Rochester.
}
